+++ NEWS.md
@@ -4,7 +4,7 @@ Red Panda C++ Version 1.0.2
   - enhancement: better display when input with IM under column mode
   - enhancement: better display current lines under column mode
   - change: test to use utf-8 as the default encoding (prepare to use libclang to implement parser)
-  - fix: auto syntax check use wrong charset, if a file in editing is not encoded with ANSI encoding
+  - fix: auto syntax check fail, if the file is not gbk and includes files encoded with utf8
   - enhancement: timeout for problem case test
   - enhancement: slightly reduce start up time
   - enhancement: use icon to indicate missing project files in the project view